Changes to be committed: 	modified:   src/org/openscenegraph/osg/viewer/OSGRenderer.java 	new file:   src/org/openscenegraph/osg/viewer/OffScreenViewer.java 	modified:   src/org/openscenegraph/osg/viewer/Viewer.java 	new file:   src/org/openscenegraph/osg/viewer/ViewerBase.java
Added newly JNI functions and classes to allow for off-screen rendering
on Android.
The viewer, formerly just being JNIviewer, is now separated into Viewer
and OffScreenViewer, which both inherit from ViewerBase.

NOTE: just ran an ndk build - errors came up as the normally distributed
OSG version apparently doesn't have a RefVec2 (just the way it provides
one for Vec3 and Vec4). I know I added it locally to an OSG build (only
thing I changed is defining the RefVec2) and submitted it to osg-submissions
mailing list. please check in the osg trunk of 3.4 or later if it's there
for you - OR: add it as external definition is the JNI headers here. Don't
know what's more handy.